requirements.txt
# 核心框架
flask==2.3.3
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0

# 数据处理
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
scipy==1.11.4

# 图数据库
neo4j==5.14.1
py2neo==2021.2.4

# 自然语言处理
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
jieba==0.42.1
openai==1.3.7

# 网络爬虫
requests==2.31.0
beautifulsoup4==4.12.2
scrapy==2.11.0
selenium==4.15.2

# 数据库
redis==5.0.1
mysql-connector-python==8.2.0
PyMySQL==1.1.0
sqlalchemy==2.0.23

# 机器学习
lightgbm==4.1.0
xgboost==2.0.2
catboost==1.2.2

# 可视化
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# API和Web
pydantic==2.5.0
jinja2==3.1.2
celery==5.3.4

# 工具库
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1
loguru==0.7.2

# 测试
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# 开发工具
black==23.11.0
flake8==6.1.0
mypy==1.7.1